Remove type duplication
