+++ packages/libs/core/src/types.ts
@@ -103,6 +103,12 @@ export type PageManifest = Manifest & {
   trailingSlash: boolean;
 };
 
+export type HeaderData = {
+  source: string;
+  headers: Header[];
+  regex: string;
+};
+
 export type I18nData = {
   locales: string[];
   defaultLocale: string;
@@ -112,6 +118,7 @@ export type RoutesManifest = {
   basePath: string;
   redirects: RedirectData[];
   rewrites: RewriteData[];
+  headers: HeaderData[];
   i18n?: I18nData;
 };
