+++ packages/libs/lambda-at-edge/src/types.ts
@@ -4,27 +4,7 @@ import type {
   CloudFrontResponse
 } from "aws-lambda";
 import { ApiManifest, PageManifest } from "@sls-next/core";
-
-export type DynamicPageKeyValue = {
-  [key: string]: {
-    file: string;
-    regex: string;
-  };
-};
-
-// Image optimization
-export type ImageConfig = {
-  deviceSizes: number[];
-  imageSizes: number[];
-  loader: "default" | "imgix" | "cloudinary" | "akamai";
-  path: string;
-  domains?: string[];
-};
-
-export type ImagesManifest = {
-  version: number;
-  images: ImageConfig;
-};
+export { ImageConfig, ImagesManifest, RoutesManifest } from "@sls-next/core";
 
 export type OriginRequestApiHandlerManifest = ApiManifest & {
   enableHTTPCompression?: boolean;
@@ -82,44 +62,6 @@ export type PreRenderedManifest = {
     previewModeSigningKey: string;
     previewModeEncryptionKey: string;
   };
-};
-
-export type RedirectData = {
-  statusCode: number;
-  source: string;
-  destination: string;
-  regex: string;
-  internal?: boolean;
-};
-
-export type RewriteData = {
-  source: string;
-  destination: string;
-  regex: string;
-};
-
-export type Header = {
-  key: string;
-  value: string;
-};
-
-export type HeaderData = {
-  source: string;
-  headers: Header[];
-  regex: string;
-};
-
-export type I18nData = {
-  locales: string[];
-  defaultLocale: string;
-};
-
-export type RoutesManifest = {
-  basePath: string;
-  redirects: RedirectData[];
-  rewrites: RewriteData[];
-  headers: HeaderData[];
-  i18n?: I18nData;
 };
 
 export type PerfLogger = {
